+++ docs/src/examples/core/BasicTree.vue
@@ -20,7 +20,7 @@ const { items } = api
         :class="cn('min-h-8 px-1 flex items-center gap-1 rounded relative dark:hover:bg-zinc-800 hover:bg-zinc-100', {
           'ring-[1px] ring-primary': item.isSelected(),
         })"
-        @click="api.onClick(item.id, $event)"
+        @click="api.inputNodeClick(item.id, $event)"
       >
         <button
           :class="cn('ml-1 transition-all rounded-full size-5 flex items-center justify-center hover:bg-zinc-200', { 'opacity-0': !item.hasChildren() })"