+++ docs/src/examples/keyboard/BasicNavigation.vue
@@ -8,7 +8,7 @@ import { Kbd, KbdGroup } from '@/components/ui/kbd'
 import { treeData } from '@/utils/data'
 import { cn } from '@/utils/shadcn'
 
-const { items, onClick } = useSpliceTree(treeData, {
+const { items, inputNodeClick } = useSpliceTree(treeData, {
   plugins: [keyboard, pointer, selectable],
   configuration: {
     keyboard: {
@@ -37,7 +37,7 @@ const { items, onClick } = useSpliceTree(treeData, {
         :class="cn('min-h-8 px-1 flex items-center gap-1 rounded relative dark:hover:bg-zinc-800 hover:bg-zinc-100', {
           'ring-[1px] ring-primary': item.isSelected(),
         })"
-        @click="onClick(item.id, $event)"
+        @click="inputNodeClick(item.id, $event)"
       >
         <button
           :class="cn('ml-1 transition-all rounded-full size-5 flex items-center justify-center hover:bg-zinc-200', { 'opacity-0': !item.hasChildren() })"