+++ packages/plugins/lazy-load/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-lazy-load
 
+## 1.0.0
+
 ## 0.3.0
 
 ## 0.2.0