+++ packages/plugins/search/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-search
 
+## 1.0.0
+
 ## 0.3.0
 
 ## 0.2.0